feat: 지하철 역 조회 기능

+++ src/main/java/subway/service/StationService.java
@@ -1,5 +1,7 @@
 package subway.service;
 
+import java.util.List;
+
 import subway.domain.Station;
 import subway.domain.StationRepository;
 import subway.view.StationView;
@@ -14,11 +16,16 @@ public class StationService {
 	}
 
 	public static boolean deleteStation(String stationName) {
-		if (containsStation(stationName)) {
+		if (containsStation(stationName) && LineService.isFirstOrEnd(findStationById(stationName))) {
 			StationRepository.deleteStation(stationName);
 			return true;
 		}
 		return false;
+	}
+
+	public static List<Station> getAllStations() {
+		List<Station> stationList = StationRepository.stations();
+		return stationList;
 	}
 
 	private static boolean containsStation(String name) {
@@ -29,5 +36,10 @@ public class StationService {
 			return true;
 		}
 		return false;
+	}
+
+	private static Station findStationById(String name) {
+		return StationRepository.stations().stream()
+			.filter(station -> station.getName().equals(name)).findFirst().get();
 	}
 }
